hf_space/requirements.txt
gradio
httpx
torch
torchvision
pillow
numpy

